+++ src/main.js
@@ -4,3 +4,4 @@ import "./variations/classics.js";
 import "./variations/exotic.js";
 import "./variations/solarized.js";
 import "./variations/stained.js";
+import "./variations/surprising.js";